requirements.txt
faiss==1.9.0
kagglehub==0.3.13
numpy==2.3.5
PyYAML==6.0.3
rank_bm25==0.2.2
torch==2.2.2
torch==2.9.1+cu128
transformers==5.0.0.dev0
